fastapi-backend/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
pydantic==2.12.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
aiomysql==0.3.2
PyMySQL==1.1.2
Authlib==1.6.6
PyJWT==2.10.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.3
itsdangerous==2.2.0
redis==7.1.0
httpx==0.28.1
requests==2.32.5